backend/requirements.txt
flask==0.12.2
sqlalchemy==1.1.11
psycopg2==2.7.1
flask-migrate==2.1.0
flask-script==2.0.5
flask-cors==3.0.3
flask-marshmallow==0.8.0
marshmallow-sqlalchemy==0.13.1
flask-sso==0.4.0
pyjwt==1.5.3
flask-httpauth==3.2.3
elasticsearch_dsl==6.1.0
flask-uploads==0.2.1
python-magic==0.4.13
requests-toolbelt==0.8.0
flask-mail==0.9.1
icalendar==4.0.0
markdown2==2.3.4
flask-compress==1.4.0
boto3==1.7.40